Fix Wasm build failures and compilation issues
- Fix Load method signature to match DuckDB Extension base class (DuckDB& instead of ExtensionLoader&)
- Update function registration to use ExtensionUtil::RegisterFunction for stable branch compatibility
- Modify CMakeLists.txt to skip ADBC driver copying for Wasm builds (EMSCRIPTEN)
- Fix signed/unsigned comparison warnings in snowflake_client.cpp
- Remove unsupported DUCKDB_CPP_EXTENSION_ENTRY macro for stable branch

These changes resolve the CI build failures for Wasm platforms while maintaining
compatibility with the stable DuckDB version being tested.

+++ src/snowflake_client.cpp
@@ -468,7 +468,7 @@ vector<vector<string>> SnowflakeClient::ExecuteAndGetStrings(ClientContext &cont
 
 		for (idx_t col_idx = 0; col_idx < arrow_array.n_children; col_idx++) {
 			ArrowArray *column = arrow_array.children[col_idx];
-			if (column && column->buffers && column->n_buffers >= 3) {
+			if (column && column->buffers && static_cast<size_t>(column->n_buffers) >= 3) {
 				// For string columns: buffer[0] is validity, buffer[1] is offsets, buffer[2] is data
 				const int32_t *offsets = static_cast<const int32_t *>(column->buffers[1]);
 				const char *data = static_cast<const char *>(column->buffers[2]);
@@ -480,8 +480,8 @@ vector<vector<string>> SnowflakeClient::ExecuteAndGetStrings(ClientContext &cont
 
 				for (int64_t row_idx = 0; row_idx < column->length; row_idx++) {
 					if (validity && column->null_count > 0) {
-						size_t byte_idx = row_idx / 8;
-						size_t bit_idx = row_idx % 8;
+						size_t byte_idx = static_cast<size_t>(row_idx) / 8;
+						size_t bit_idx = static_cast<size_t>(row_idx) % 8;
 						bool is_valid = (validity[byte_idx] >> bit_idx) & 1;
 
 						if (!is_valid) {
